Implemented functionality for manager article editing
1. Changed the look of article list into bootstrap cards.
2. If the manager is logged in, in the top right of every article there is a button for editing which opens dialog form.
3. If the user makes changes to an article but closes the dialog before saving changes, the form is reset to the values before the changes.

+++ FoodDelivery/src/beans/Article.java
@@ -7,11 +7,11 @@ public class Article {
 	private double price;
 	private ArticleType type;
 	private String restaurantName;
-	private double quantity;
+	private int quantity;
 	private String description;
 	private String image;
 	
-	public Article(String name, double price, ArticleType type, String restaurantName, double quantity,
+	public Article(String name, double price, ArticleType type, String restaurantName, int quantity,
 			String description, String image) {
 		super();
 		this.name = name;
@@ -59,11 +59,11 @@ public class Article {
 		this.restaurantName = restaurantName;
 	}
 
-	public double getQuantity() {
+	public int getQuantity() {
 		return quantity;
 	}
 
-	public void setQuantity(double quantity) {
+	public void setQuantity(int quantity) {
 		this.quantity = quantity;
 	}
